Change structure of 2 minute trigger handler
The script is running every 10 minutes (not 2) 

The ncrontab expression checker indicates that this should run on the exact 2 minute mark (which makes sense & is a nice reliable time)

However, this set of docs from Azure doesn't have any examples of this structure and does use the asterisk in their "run every 5 minutes example" - wondering if there is an issue with it. Probably not - but testing to see.